feat!: implement `exitEntriesKeys` & `sleepBeforeNextEntry()` to ensure messages sequential rendering
BREAKING CHANGE: Before this change, we could have race-conditions where the client asked the library to display multiple messages at once. The goal of this commit was to ensure that all `setNewEntry()` calls are now queued and processed one at a time. The breaking change here is that it's now the responsibility of the client to ensure that `sleepBeforeNextEntry()` is called before any subsequent call of `setNewEntry()`.